requirements.txt
openai
line-bot-sdk
langchain
google-api-python-client
requests
